Initial commit for the D3D12 backed Gorgon.Graphics.Core
There is an absolutely massive amount of code in here, much of it is work in progress and will almost certainly change over the next year or two. Only pull this branch if you're feeling brave and/or stupid and won't bug me about why things are the way they are or bugs.

Thou hath been warned...eth.

+++ Gorgon/Gorgon.Input/_Internal/InputWindow.cs
@@ -115,7 +115,7 @@ internal sealed class InputWindow
                 }
                 return (LRESULT)0;
             case PInvoke.WM_INPUT_DEVICE_CHANGE:
-                switch ((uint)wParam)
+                switch ((nuint)wParam)
                 {
                     case PInvoke.GIDC_ARRIVAL:
                         _deviceAttached?.Invoke(new HANDLE(lParam.Value));
